Allow rewriting client span ids and remove ssampling
Rewriting client span ids is controlled by a flag on the tracespan
reported to the adapter, which we hardcode to 'true' in our supplied
tracespan instance resource.  This will change client span ids to
another span id in a deterministic manner, as well as any parent span
ids of server spans.  This is necessary since, by default, Istio reports
a client/server span pair with the same id, which doesn't work well on
our backend.

Removed sampling because it isn't needed anymore with our gateway
architecture.